+++ src/apps/encryption/mod.rs
@@ -1,9 +1,11 @@
 mod age;
 mod doppler;
+mod gocryptfs;
 mod gopass;
 mod pass_cli;
 
 pub use age::Age;
 pub use doppler::Doppler;
+pub use gocryptfs::Gocryptfs;
 pub use gopass::Gopass;
 pub use pass_cli::PassCli;